pyproject.toml
[project]
name = "quic-360"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Silviase", email = "koki.maeda@nlp.c.titech.ac.jp" }
]
dependencies = [
    "datasets>=3.0.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/quic_360"]
